fix: pod restart on config change
In order to restart the pods on any config change, a hash of the config is added as an annotation to the deployment. Additionally the validator config and policy are merged together, so that both reside in a ConfigMap.

fixes #251
