--- connaisseur/res/policy_schema.json
@@ -1,29 +1,0 @@
-{
-    "type": "object",
-    "properties": {
-        "rules": {
-            "type": "array",
-            "items": {
-                "type": "object",
-                "properties": {
-                    "pattern": {
-                        "type": "string",
-                        "pattern": "^((\\w+\\.[\\.\\w]*\\/)?([^\\s]+\\/)?)([^\\/\\:@]+)((@sha256:([a-f0-9]{64}))|(:(.+)))?$"
-                    },
-                    "validator": {
-                        "type": "string"
-                    },
-                    "with": {
-                        "type": "object"
-                    }
-                },
-                "required": [
-                    "pattern"
-                ]
-            }
-        }
-    },
-    "required": [
-        "rules"
-    ]
-}
