--- helm/crds/imagepolicy.yaml
@@ -1,51 +1,0 @@
-apiVersion: apiextensions.k8s.io/v1
-kind: CustomResourceDefinition
-metadata:
-  # name must match the spec fields below, and be in the form: <plural>.<group>
-  name: imagepolicies.connaisseur.policy
-  labels:
-    app.kubernetes.io/instance: connaisseur
-spec:
-  # group name to use for REST API: /apis/<group>/<version>
-  group: connaisseur.policy
-  # list of versions supported by this CustomResourceDefinition
-  versions:
-    - name: v1
-      # Each version can be enabled/disabled by Served flag.
-      served: true
-      # One and only one version must be marked as the storage version.
-      storage: true
-      schema:
-        openAPIV3Schema:
-          type: object
-          properties:
-            spec:
-              type: object
-              properties:
-                rules:
-                  type: array
-                  items:
-                    type: object
-                    properties:
-                      pattern:
-                        type: string
-                        pattern: '^((\w+\.[\.\w]*\/)?([^\s]+\/)?)([^\/\:@]+)((@sha256:([a-f0-9]{64}))|(:(.+)))?$'
-                      validator:
-                        type: string
-                      with:
-                        type: object
-                        additionalProperties: true
-                    required:
-                      - pattern
-  # either Namespaced or Cluster
-  scope: Cluster
-  names:
-    # plural name to be used in the URL: /apis/<group>/<version>/<plural>
-    plural: imagepolicies
-    # singular name to be used as an alias on the CLI and for display
-    singular: imagepolicy
-    # kind is normally the CamelCased singular type. Your resource manifests use this.
-    kind: ImagePolicy
-    # shortNames allow shorter string to match your resource on the CLI
-    shortNames:
-      - imgpol
