+++ helm/templates/env.yaml
@@ -7,7 +7,6 @@ metadata:
     {{- include "helm.labels" . | nindent 4 }}
 data:
   CONNAISSEUR_NAMESPACE: {{ .Release.Namespace }}
-  CONNAISSEUR_IMAGE_POLICY: {{ .Chart.Name }}-policy
   KUBE_API_TOKEN_PATH: /var/run/secrets/kubernetes.io/serviceaccount/token
   KUBE_API_CA_PATH: /var/run/secrets/kubernetes.io/serviceaccount/ca.crt
   {{- if .Values.debug }}
