--- helm/templates/policy.yaml
@@ -1,11 +1,0 @@
-apiVersion: "connaisseur.policy/v1"
-kind: ImagePolicy
-metadata:
-  name: {{ .Chart.Name }}-policy
-  labels:
-    {{- include "helm.labels" . | nindent 4 }}
-spec:
-  rules:
-    {{- range $k,$v := .Values.policy }}
-     - {{ toJson $v }}
-    {{- end }}
