+++ mkdocs.yml
@@ -105,6 +105,7 @@ nav:
       - adr/ADR-3_multi_notary_config.md
       - adr/ADR-4_modular.md
       - adr/ADR-5_no-more-bootstrap.md
+      - adr/ADR-6_dynamic-config.md
   - Releases ⧉: https://github.com/sse-secure-systems/connaisseur/releases
   - Support ⧉: https://github.com/sse-secure-systems/connaisseur/discussions
