+++ tests/data/config/err4.yaml
@@ -1,3 +1,4 @@
+validators:
 - name: default
   host: notary.docker.io
   trust_roots:
@@ -45,3 +46,31 @@
       -----END PUBLIC KEY-----
   auth:
     secretName: harbor-creds
+policy:
+- pattern: "*:*"
+  with:
+    delegations: ["phbelitz", "chamsen"]
+- pattern: "docker.io/*:*"
+  validator: dockerhub
+  with:
+    delegations: ["phbelitz"]
+- pattern: "k8s.gcr.io/*:*"
+  validator: allow
+- pattern: "gcr.io/*:*"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/*:*"
+  validator: dockerhub
+  with:
+    delegations: ["someuserthatdidnotsign"]
+- pattern: "docker.io/securesystemsengineering/sample"
+  validator: dockerhub
+  with:
+    delegations: ["phbelitz", "chamsen"]
+- pattern: "docker.io/securesystemsengineering/sample:v4"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/connaisseur:*"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/sample-san-sama"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/alice-image"
+  validator: dockerhub
