+++ tests/data/config/sample_config.yaml
@@ -1,3 +1,4 @@
+validators:
 - name: default
   type: notaryv1
   host: notary.docker.io
@@ -70,3 +71,31 @@
         MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEOXYta5TgdCwXTCnLU09W5T4M4r9f
         QQrqJuADP6U7g5r9ICgPSmZuRHP/1AYUfOQW3baveKsT969EfELKj1lfCA==
         -----END PUBLIC KEY-----
+policy:
+- pattern: "*:*"
+  with:
+    delegations: ["phbelitz", "chamsen"]
+- pattern: "docker.io/*:*"
+  validator: dockerhub
+  with:
+    delegations: ["phbelitz"]
+- pattern: "k8s.gcr.io/*:*"
+  validator: allow
+- pattern: "gcr.io/*:*"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/*:*"
+  validator: dockerhub
+  with:
+    delegations: ["someuserthatdidnotsign"]
+- pattern: "docker.io/securesystemsengineering/sample"
+  validator: dockerhub
+  with:
+    delegations: ["phbelitz", "chamsen"]
+- pattern: "docker.io/securesystemsengineering/sample:v4"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/connaisseur:*"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/sample-san-sama"
+  validator: allow
+- pattern: "docker.io/securesystemsengineering/alice-image"
+  validator: dockerhub
